+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 
 ### Fixes
 
+- Fixed wrong number formatting on staking screens ([PR 2213](https://github.com/input-output-hk/daedalus/pull/2213))
 - Removed broken feature requests link ([PR 2212](https://github.com/input-output-hk/daedalus/pull/2212))
 - Fixed animation on wallet dropdown option during redemption process ([PR 2191](https://github.com/input-output-hk/daedalus/pull/2191))
 - Fixed Cardano Explorer URLs ([PR 2198](https://github.com/input-output-hk/daedalus/pull/2198))